+++ src/stats.rs
@@ -177,6 +177,15 @@ impl GooseStats {
                 fmt,
                 " ------------------------+----------------+----------------+--------+--------- "
             )?;
+            let req_s;
+            let fail_s;
+            if self.duration == 0 {
+                req_s = 0.to_formatted_string(&Locale::en);
+                fail_s = 0.to_formatted_string(&Locale::en);
+            } else {
+                req_s = (aggregate_total_count / self.duration).to_formatted_string(&Locale::en);
+                fail_s = (aggregate_fail_count / self.duration).to_formatted_string(&Locale::en);
+            }
             // Compress 100.0 and 0.0 to 100 and 0 respectively to save width.
             if aggregate_fail_percent as usize == 100 || aggregate_fail_percent as usize == 0 {
                 writeln!(
@@ -189,8 +198,8 @@ impl GooseStats {
                         aggregate_fail_count.to_formatted_string(&Locale::en),
                         aggregate_fail_percent as usize
                     ),
-                    (aggregate_total_count / self.duration).to_formatted_string(&Locale::en),
-                    (aggregate_fail_count / self.duration).to_formatted_string(&Locale::en),
+                    req_s,
+                    fail_s,
                 )?;
             } else {
                 writeln!(
@@ -203,8 +212,8 @@ impl GooseStats {
                         aggregate_fail_count.to_formatted_string(&Locale::en),
                         aggregate_fail_percent
                     ),
-                    (aggregate_total_count / self.duration).to_formatted_string(&Locale::en),
-                    (aggregate_fail_count / self.duration).to_formatted_string(&Locale::en),
+                    req_s,
+                    fail_s,
                 )?;
             }
         }